// assign/src/main/angularNew/src/app/model/e-book.ts
export class EBook {

    id: string;
    name: string;
    rating: string;
    email: string;

    constructor(){}

    // constructor(id: string, name: string, rating: string, email: string){
    //     this.id = id;
    //     this.name = name;
    //     this.rating = rating;
    //     this.email = email;
    // }
}